+++ actix-utils/CHANGES.md
@@ -1,5 +1,12 @@
 # Changes
 
+## [0.2.2] - 2019-02-11
+
+### Added
+
+* Add `Display` impl for `TimeoutError`
+
+
 ## [0.2.1] - 2019-02-06
 
 ### Added